Topic/units

* Changes to C and C++ interface calls that make string sizes explicit.
* Stricter enforcement of direct type mapping (e.g. Python int is not compatible with int32 because default is int64 and precision could be lost).
* Creation of schemas for validating YAMLs and metaschema for validating datatype schemas.
* Addition of datatype option in YAML.
* Support user defined data types via JSON schema.
* Changes to the internal structure of Ply and Obj data structures.
* Tables are no longer automatically consolidated into structure numpy arrays in Python/Matlab on recv. Instead, use 'recv_array' to do so.
* Removed generic RPComm. Instead use the ServerComm or ClientComm
* Removed deprecated functions beginning with "Psi"
* Removed send_line/recv_line methods (these were aliases for regular send/recv)
* Removed send_row/recv_row methods (these were aliases for regular send/recv)

+++ cis_interface/examples/rpcFib/src/rpcFibSrv.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
 import sys
+import numpy as np
 from cis_interface.interface.CisInterface import CisRpcServer
 from cis_interface.tools import sleep
 
@@ -37,7 +38,7 @@ def fibServer(args):
 
         # Sleep and then send response back
         sleep(float(sleeptime))
-        flag = rpc.rpcSend(arg, result)
+        flag = rpc.rpcSend(arg, np.int32(result))
         if not flag:
             raise RuntimeError('rpcFibSrv(P): ERROR sending')
