+++ cis_interface/examples/rpc_lesson1/src/client.py
@@ -1,4 +1,5 @@
 import sys
+import numpy as np
 from cis_interface.interface.CisInterface import (
     CisRpcClient, CisOutput)
 
@@ -24,14 +25,14 @@ def main(iterations):
         
         # Call the server and receive response
         print('client(Python): Calling fib(%d)' % i)
-        ret, result = rpc.call(i)
+        ret, result = rpc.call(np.int32(i))
         if not ret:
             raise RuntimeError('client(Python): RPC CALL ERROR')
         fib = result[0]
         print('client(Python): Response fib(%d) = %d' % (i, fib))
 
         # Log result by sending it to the log connection
-        ret = log.send(i, fib)
+        ret = log.send(np.int32(i), fib)
         if not ret:
             raise RuntimeError('client(Python): SEND ERROR')
